owner id int로 되돌림

+++ src/main/java/kr/co/solfood/owner/store/OwnerStoreController.java
@@ -28,7 +28,7 @@ public class OwnerStoreController {
         Optional<OwnerVO> ownerOpt = Optional.ofNullable((OwnerVO) sess.getAttribute("ownerLoginSession"));
 
         if (ownerOpt.isPresent()) {
-            long ownerId = ownerOpt.get().getOwnerId();
+            int ownerId = ownerOpt.get().getOwnerId();
             OwnerStoreVO store = ownerStoreService.getOwnerStore(ownerId);
             model.addAttribute("store", store);
             return "owner/store";
